requirements.txt
openai-whisper
youtube_transcript_api~=0.6.2
pytube~=15.0.0
openai~=1.25.1
telebot~=0.0.5